requirements.txt
asn1crypto==0.24.0
async-generator==1.10
attrs==21.4.0
Automat==0.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.12
click==6.7
cloud-init==20.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
distro-info===0.18ubuntu0.18.04.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
filelock==3.7.0
fonttools==4.31.2
gdown==4.4.0
h11==0.13.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
imageio==2.19.0
importlib==1.0.4
importlib-metadata==4.11.4
incremental==16.10.1
ipykernel==6.9.2
ipynb==0.5.1
ipython==7.32.0
jedi==0.18.1
Jinja2==2.10
joblib==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kaggle==1.5.12
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.0
language-selector==0.1
MarkupSafe==1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
numpy==1.21.5
oauthlib==2.0.6
opencv-python==4.5.5.64
outcome==1.1.0
packaging==21.3
PAM==0.4.2
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.11.2
PyGObject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==3.0.7
pyserial==3.4
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
python-debian==0.1.32
python-slugify==6.1.1
pytz==2022.1
pyxdg==0.25
PyYAML==3.12
pyzmq==22.3.0
requests==2.28.0
requests-unixsocket==0.1.5
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
SecretStorage==2.3.1
selenium==4.1.3
service-identity==16.0.0
six==1.11.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
split-folders==0.5.1
ssh-import-id==5.7
systemd-python==234
text-unidecode==1.3
threadpoolctl==3.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchsummary==1.5.1
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==17.9.0
typing_extensions==4.1.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
wcwidth==0.2.5
webdriver-manager==3.5.4
wsproto==1.1.0
zipp==3.8.0
zope.interface==4.3.2
